# Conflicts:
#	server/node/src/main/java/org/kaaproject/kaa/server/admin/services/KaaAdminServiceImpl.java

+++ server/node/src/main/java/org/kaaproject/kaa/server/admin/services/KaaAdminServiceImpl.java
@@ -354,9 +354,7 @@ public class KaaAdminServiceImpl implements KaaAdminService, InitializingBean {
         checkAuthority(KaaAuthorityDto.TENANT_DEVELOPER, KaaAuthorityDto.TENANT_USER);
         try {
             EndpointProfileBodyDto profileBodyDto = controlService.getEndpointProfileBodyByKeyHash(endpointProfileKeyHash);
-            if (profileBodyDto == null) {
-                throw new KaaAdminServiceException("Requested item was not found!", ServiceErrorCode.ITEM_NOT_FOUND);
-            }
+            Utils.checkNotNull(profileBodyDto);
             checkApplicationId(profileBodyDto.getAppId());
             return profileBodyDto;
         } catch (Exception e) {
@@ -371,9 +369,11 @@ public class KaaAdminServiceImpl implements KaaAdminService, InitializingBean {
         checkAuthority(KaaAuthorityDto.TENANT_DEVELOPER, KaaAuthorityDto.TENANT_USER);
         try {
             EndpointProfileDto profileDto = controlService.getEndpointProfileByKeyHash(endpointKeyHash);
+            Utils.checkNotNull(profileDto);
             checkApplicationId(profileDto.getApplicationId());
             ServerProfileSchemaDto serverProfileSchema = controlService.getServerProfileSchemaByApplicationIdAndVersion(
                     profileDto.getApplicationId(), serverProfileVersion);
+            Utils.checkNotNull(serverProfileSchema);
             RecordField record;
             try {
                 record = createRecordFieldFromCtlSchemaAndBody(serverProfileSchema.getCtlSchemaId(),
@@ -653,6 +653,9 @@ public class KaaAdminServiceImpl implements KaaAdminService, InitializingBean {
             UserDto user = controlService.getUser(userId);
             Utils.checkNotNull(user);
             checkTenantId(user.getTenantId());
+            if (KaaAuthorityDto.KAA_ADMIN.equals(user.getAuthority()) || KaaAuthorityDto.TENANT_ADMIN.equals(user.getAuthority())) {
+                throw new KaaAdminServiceException("Can't delete KAA admin or Tenant admin user!", ServiceErrorCode.PERMISSION_DENIED);
+            }
             userFacade.deleteUser(Long.valueOf(user.getExternalUid()));
             controlService.deleteUser(user.getId());
         } catch (Exception e) {
@@ -1751,7 +1754,6 @@ public class KaaAdminServiceImpl implements KaaAdminService, InitializingBean {
         try {
             checkEndpointGroupId(endpointGroupId);
             ProfileFilterRecordDto record = controlService.getProfileFilterRecord(endpointProfileSchemaId, serverProfileSchemaId, endpointGroupId);
-            Utils.checkNotNull(record);
             return record;
         } catch (Exception e) {
             throw Utils.handleException(e);
@@ -1870,7 +1872,6 @@ public class KaaAdminServiceImpl implements KaaAdminService, InitializingBean {
         try {
             ProfileFilterRecordDto record = controlService.getProfileFilterRecord(endpointProfileSchemaId,
                     serverProfileSchemaId, endpointGroupId);
-            Utils.checkNotNull(record);
             checkEndpointGroupId(record.getEndpointGroupId());
             String username = getCurrentUser().getUsername();
             controlService.deleteProfileFilterRecord(endpointProfileSchemaId, serverProfileSchemaId, endpointGroupId, username);
@@ -1898,7 +1899,6 @@ public class KaaAdminServiceImpl implements KaaAdminService, InitializingBean {
         try {
             checkEndpointGroupId(endpointGroupId);
             ConfigurationRecordDto record = controlService.getConfigurationRecord(schemaId, endpointGroupId);
-            Utils.checkNotNull(record);
             return record;
         } catch (Exception e) {
             throw Utils.handleException(e);
@@ -2119,7 +2119,6 @@ public class KaaAdminServiceImpl implements KaaAdminService, InitializingBean {
         checkAuthority(KaaAuthorityDto.TENANT_DEVELOPER, KaaAuthorityDto.TENANT_USER);
         try {
             StructureRecordDto<ConfigurationDto> record = controlService.getConfigurationRecord(schemaId, endpointGroupId);
-            Utils.checkNotNull(record);
             checkEndpointGroupId(record.getEndpointGroupId());
             String username = getCurrentUser().getUsername();
             controlService.deleteConfigurationRecord(schemaId, endpointGroupId, username);
